app/requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
h5py==3.13.0
imbalanced-learn==0.13.0
imblearn==0.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
numpy==2.2.6
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sklearn-compat==0.1.3
threadpoolctl==3.6.0
tzdata==2025.2
Werkzeug==3.1.3
